requirements.txt
matplotlib==3.7.5
numpy==1.24.4
opencv_python==4.10.0.84
Pillow==11.0.0
scikit_learn==1.5.2
torch==2.5.0
torchvision==0.20.0
tqdm==4.66.5
